requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-cloud-secret-manager>=2.16.0
python-dotenv>=1.0.0
